The code has been refactored and now contains only the code for the biaffine model. Simplified the multiplication operations in the biaffine model.
